Refactor to remove Django template inheritance
