+++ app/server/static/js/annotation.pug
@@ -1,0 +1,119 @@
+div.columns(v-cloak="")
+  aside.column.is-3.aside.hero.is-fullheight
+    div
+
+      div.main.pr20.pl20
+        div.field.has-addons
+          div.control.is-expanded
+            input.input(
+              type="text",
+              placeholder="Search document",
+              v-model="searchQuery",
+              @keyup.enter="submit",
+              style="border-right: none; box-shadow: none; -webkit-box-shadow: none;"
+            )
+
+          div.control
+            div.dropdown.is-hoverable.is-right
+              div.dropdown-trigger
+                button.button(
+                  aria-haspopup="true",
+                  aria-controls="dropdown-menu",
+                  style="border-left: none"
+                )
+                  span.icon.has-text-grey.pr0
+                    i.fas.fa-angle-down(aria-hidden="true")
+
+              div#dropdown-menu.dropdown-menu.pt0(role="menu")
+                div.dropdown-content
+                  a.dropdown-item
+                    label.radio
+                      input(
+                        type="radio",
+                        value="all",
+                        v-model="picked",
+                        checked=""
+                      )
+                      | All
+                  a.dropdown-item
+                    label.radio
+                      input(
+                        type="radio",
+                        value="active",
+                        v-model="picked"
+                      )
+                      | Active
+                  a.dropdown-item
+                    label.radio
+                      input(
+                        type="radio",
+                        value="completed",
+                        v-model="picked"
+                      )
+                      | Completed
+
+      div.main.pt0.pb0.pr20.pl20
+        span About {{ count }} results
+
+      div.main
+        a.item(
+          href="#",
+          v-for="(doc, index) in docs",
+          :class="{ active: index == pageNumber }",
+          @click="pageNumber = index",
+          :data-preview-id="index"
+        )
+          span.icon
+            i.fa.fa-check(v-show="annotations[index] && annotations[index].length")
+          span.name {{ doc.text.slice(0, 60) }}...
+
+  div#message-pane.column.is-7.is-offset-1.message.hero.is-fullheight
+
+    div.modal(:class="{ 'is-active': isActive }")
+        div.modal-background
+        div.modal-card
+          header.modal-card-head
+            p.modal-card-title Annotation Guideline
+            button.delete(aria-label="close", @click="isActive = !isActive")
+          section.modal-card-body.modal-card-body-footer.content(
+            style="line-height:150%",
+            v-html="compiledMarkdown"
+          )
+
+    div.columns.is-multiline.is-gapless.is-mobile.is-vertical-center
+      div.column.is-3
+        progress.progress.is-inline-block(
+          :class="progressColor",
+          :value="achievement",
+          max="100"
+        ) 30%
+      div.column.is-8
+        span.ml10
+          strong {{ total - remaining }}
+        | /
+        span {{ total }}
+      div.column.is-1.has-text-right
+        a.button(@click="isActive =! isActive")
+          span.icon
+            i.fas.fa-book
+
+    block annotation-area
+
+    div.level.mt30
+      a.button.level-left(
+        @click="prevPage",
+        v-shortkey="{prev1: ['ctrl', 'p'], prev2: ['arrowup'], prev3: ['arrowleft']}",
+        @shortkey="prevPage"
+      )
+        span.icon
+          i.fas.fa-chevron-left
+        span Prev
+
+      a.button.level-right(
+        @click="nextPage",
+        v-shortkey="{next1: ['ctrl', 'n'], next2: ['arrowdown'], next3: ['arrowright']}",
+        @shortkey="nextPage"
+      )
+        span Next
+        span.icon
+          i.fas.fa-chevron-right
