+++ app/server/static/js/directives.js
@@ -1,0 +1,5 @@
+export default function todoFocus(el, binding) {
+  if (binding.value) {
+    el.focus();
+  }
+}
