+++ app/server/webpack.config.js
@@ -35,6 +35,10 @@ module.exports = {
     module: {
         rules: [
             {
+                test: /\.pug$/,
+                loader: 'pug-plain-loader'
+            },
+            {
                 test: /\.vue$/,
                 loader: 'vue-loader'
             }
